# src/main/resources/swagger.yml
swagger: '2.0'
info:
  title: SpongeblogSP API
  description: Spongeblog blogging API
  version: "1.0.0"
# array of all schemes that your API supportsf
schemes:
  - https
# will be prefixed to all paths
basePath: /api/v1
consumes:
  - application/json
produces:
  - application/json
parameters:
...
  IdCategory:
    name: category-id
    in: path
    description: Unique identifier of a Category;
    required: true
    type: integer
    format: int64
...
  CategoryEdit:
    name: category-edit
    in: body
    schema:
      $ref: '#/definitions/CategoryDTO'
    required: true
...
  IdParentCategory:
    name: parent-category-id
    in: path
    description: Unique identifier of the parent Category;
    required: true
    type: integer
    format: int64
...
paths:
...
  /api/v1/categories:
    get:
      summary: All Categorys
      description: Returns all Categorys.
      tags:
        - Categories
      responses:
        200:
          description: Categories.
          schema:
            type: array
            items:
              $ref: '#/definitions/CategoryDTO'
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
    post:
      summary: Create Category
      description: Creates new Category entry.
      parameters:
        - $ref: '#/parameters/CategoryEdit'
      tags: 
        - Categories
      responses:
        201:
          description: Created Category.
          schema:
            $ref: '#/definitions/CategoryDTO'
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
    put:
      summary: Update Category
      description: Updated a Category entry.
      parameters:
        - $ref: '#/parameters/CategoryEdit'
      tags: 
        - Categories
      responses:
        200:
          description: Category updated.
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
  /api/v1/categories/{category-id}:
    get:
      summary: Get Category
      description: Fetches Category with defined Id.
      parameters:
        - $ref: '#/parameters/IdCategory'
      tags: 
        - Categories
      responses:
        200:
          description: Category with id.
          schema:
            $ref: '#/definitions/CategoryPostsDTO'
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
    delete:
      summary: Delete Category
      description: Deletes a Category with defined Id.
      parameters:
        - $ref: '#/parameters/IdCategory'
      tags: 
        - Categories
      responses:
        200:
          description: Category deleted.
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
  /api/v1/categories/children/{parent-category-id}:
    get:
      summary: Categorys of Parent.
      description: Returns all Categorys of defined parent Category.
      parameters:
        - $ref: '#/parameters/IdParentCategory'
      tags:
        - Categories
      responses:
        200:
          description: Categories.
          schema:
            type: array
            items:
              $ref: '#/definitions/CategoryPostsDTO'
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
  /api/v1/categories/rewind/{category-id}:
    put:
      summary: Rewind Category
      description: Restores previous Category entry.
      parameters:
        - $ref: '#/parameters/IdCategory'
      tags: 
        - Categories
      responses:
        201:
          description: Previous Category restored.
        400:
          description: Bad request (validation failed).
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Unauthorized (need to log in / get token).
        403:
          description: Forbidden (no rights to access resource).
        404:
          description: Entity not found.
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Request results in a conflict.
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal Server Error.
...
definitions:
...
  CategoryDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Unique identifier.
      name:
        type: string
        description: Category display name.
      parentLink:
        type: string
        description: Link to parent Category.
      parentId:
        type: integer
        format: int64
        description: Id of parent Category.
      rank:
        type: integer
        description: Determines position of Category.
  CategoryPostsDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: Unique identifier.
      name:
        type: string
        description: Category display name.
      parentLink:
        type: string
        description: Link to parent Category.
      parent:
        type: integer
        format: int64
        description: Id of parent Category.
      subCategories:
        description: Links to child Categories.
        type: array
        items:
          type: string
      posts:
        description: Link to posts within category.
        type: string
      rank:
        type: integer
        description: Determines position of Category.
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string